MovieShowings implemented
Each movie now contains a list with MovieShowings, an object containing in which cineplex and which hall the movie will be shown, at which times and its duration. It also automatically implements weekly scheduling if desired.

Every CinemaRoom instance now has a method getHallOccupation() which will return a list of MovieShowings that correspond with the given CinemaRoom.